+++ snapshot-tests-core/src/main/java/de/skuzzle/test/snapshots/impl/SnapshotTestContext.java
@@ -9,12 +9,12 @@ import java.util.stream.Stream;
 import org.apiguardian.api.API;
 import org.apiguardian.api.API.Status;
 
+import de.skuzzle.test.snapshots.ContextFiles;
 import de.skuzzle.test.snapshots.EnableSnapshotTests;
 import de.skuzzle.test.snapshots.SnapshotDsl.Snapshot;
 import de.skuzzle.test.snapshots.SnapshotTestResult;
 import de.skuzzle.test.snapshots.impl.OrphanCollectorHolder.OrphanCollector;
 import de.skuzzle.test.snapshots.io.DirectoryResolver;
-import de.skuzzle.test.snapshots.io.UncheckedIO;
 import de.skuzzle.test.snapshots.validation.Arguments;
 
 /**
@@ -156,18 +156,19 @@ public final class SnapshotTestContext {
                 .peek(collector::addPostProcessedResult)
                 .map(OrphanDetectionResult::snapshotFile)
                 .distinct()
-                .peek(orphaned -> {
+                .peek(orphanedSnapshotFile -> {
 
-                    final Path relativePath = DirectoryResolver.relativize(orphaned.getParent());
+                    final Path relativePath = DirectoryResolver.relativize(orphanedSnapshotFile.getParent());
                     if (deleteOrphaned) {
-                        UncheckedIO.delete(orphaned);
+                        final ContextFiles contextFiles = InternalSnapshotNaming.contextFilesForSnapshotFile(orphanedSnapshotFile);
+                        contextFiles.deleteFiles();
 
                         System.err.printf("Deleted orphaned snapshot file %s in %s%n",
-                                orphaned.getFileName(), relativePath);
+                                orphanedSnapshotFile.getFileName(), relativePath);
                     } else {
                         System.err.printf(
                                 "Found orphaned snapshot file. Run with '@DeleteOrphanedSnapshots' annotation to remove: %s in %s%n",
-                                orphaned.getFileName(), relativePath);
+                                orphanedSnapshotFile.getFileName(), relativePath);
                     }
                 })
                 .collect(Collectors.toList());