+++ RELEASE_NOTES.md
@@ -9,6 +9,7 @@
 
 **All Changes**: 
 
+* [#49](https://github.com/skuzzle/snapshot-tests/issues/49): Delete context files as well when deleting orphaned snapshots
 * [#57](https://github.com/skuzzle/snapshot-tests/issues/57): Support XML namespaces in XPaths when using custom comparison rules
 * Improve format and information density of XPath debugging (can be enabled with `XmlSnapshot.withEnableXPathDebugging(true)`)
 * Improve removal of internal stack frames from stacktraces of assertion errors